requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.5.3

# Environment Variables
python-dotenv==1.0.0

# Async Support
asyncio-throttle==1.0.2

# Optional: For production
gunicorn==21.2.0
